4450 - Redis/Areas/Country: Replace CountrySummary with Country
