+++ src/server/repository/redis/area/getManyCountries.ts
@@ -1,0 +1,35 @@
+import { Objects } from 'utils/objects'
+
+import { Country } from 'meta/area'
+import { Assessment, Cycle } from 'meta/assessment'
+
+import { BaseProtocol, DB } from 'server/db'
+import { CountryRepository } from 'server/repository/assessmentCycle/country'
+import { getKeyCycle, Keys } from 'server/repository/redis/keys'
+import { RedisData } from 'server/repository/redis/redisData'
+
+type Props = {
+  assessment: Assessment
+  cycle: Cycle
+  force?: boolean
+}
+
+export const getManyCountries = async (props: Props, client: BaseProtocol = DB): Promise<Array<Country>> => {
+  const { assessment, cycle, force = false } = props
+
+  const redis = RedisData.getInstance()
+  const key = getKeyCycle({ assessment, cycle, key: Keys.Area.country })
+
+  const cachedData = await redis.hgetall(key)
+  const cachedKeys = Object.keys(cachedData)
+  let countries: Array<Country>
+
+  if (Objects.isEmpty(cachedKeys) || force) {
+    countries = await CountryRepository.getMany({ assessment, cycle }, client)
+    await redis.hmset(key, ...countries.flatMap((c) => [c.countryIso, JSON.stringify(c)]))
+  } else {
+    countries = Object.values(cachedData).map((data) => JSON.parse(data))
+  }
+
+  return countries
+}
