+++ src/server/repository/redis/area/index.ts
@@ -1,7 +1,7 @@
-import { getManyCountrySummaries } from './getManyCountrySummaries'
+import { getManyCountries } from './getManyCountries'
 import { getManyRegionGroups } from './getManyRegionGroups'
 
 export const AreaRedisRepository = {
-  getManyCountrySummaries,
+  getManyCountries,
   getManyRegionGroups,
 }
